filtreleme bitti

+++ RentACar/DTOs/ArabaCreateDto.cs
@@ -4,7 +4,7 @@ public class ArabaCreateDto
 {
     public int ArabaYasi { get; set; } 
     public string ArabaAdi { get; set; } =  string.Empty;
-    public double ArabaFiyat { get; set; }
+    public decimal ArabaFiyat { get; set; }
     public float ToplamKm { get; set; }
     
     public int ModelId { get; set; }
