+++ RentACar/DTOs/ArabaResponseDto.cs
@@ -5,7 +5,7 @@ public class ArabaResponseDto
     public int ArabaId { get; set; }
     public int ArabaYasi { get; set; } 
     public string ArabaAdi { get; set; } =  string.Empty;
-    public double ArabaFiyat { get; set; }
+    public decimal ArabaFiyat { get; set; }
     public float ToplamKm { get; set; }
     
     public int MarkaId { get; set; }
